requirements.txt
streamlit
requests
Wand
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
transformers>=4.39.1
kornia
timm